Split text processors into 2 files
Hey we had to hit that 6 .cpp/.h file minimum somehow
