+++ ktextprocessors.h
@@ -1,0 +1,14 @@
+//
+// Created by kaleb on 2021-06-17.
+//
+
+#ifndef FINAL_PROJECT_KTEXTPROCESSORS_H
+#define FINAL_PROJECT_KTEXTPROCESSORS_H
+#include <string>
+
+// Kaleb Text Processors
+int count_words(const std::string text);
+int count_numbers(const std::string text);
+std::string lowercase_letters(std::string text);
+
+#endif //FINAL_PROJECT_KTEXTPROCESSORS_H
